Update README.md
This CheatSheet is mainly for White Hats or experts on the subject, and while it has some very complex points, it also has some simple ones (for people who know about the subject, of course). And no, I'm not going to talk about: DDos and social engineering, TOR/proxy, passwd spraying, SQL Injections, Linpias, XXS Cross Site and very violent techniques. This is a great help sheet for people who want to practice, study, and keep in mind this skills to used in controlled environments, not to do illegal things like stealing, scamming.
Or hacking the neighbour's wifi and turning his music off...